Split benchmark files into separate neo4j/ and doublets/ directories
Per user feedback, reorganize benchmarks to have clear separation between
Neo4j and Doublets implementations. Each file now contains only one
database's benchmarks for easy side-by-side comparison.

New structure:
- rust/benches/benchmarks/neo4j/     - All Neo4j benchmarks
- rust/benches/benchmarks/doublets/  - All Doublets benchmarks

Each directory has the same structure:
- create.rs, update.rs, delete.rs - CRUD operations
- each/all.rs, identity.rs, concrete.rs, outgoing.rs, incoming.rs - Query operations
