+++ rust/benches/benchmarks/mod.rs
@@ -3,6 +3,13 @@
 //! This module contains all benchmark implementations comparing Neo4j and Doublets.
 //! Each benchmark tests a specific database operation across all storage backends.
 //!
+//! ## Module Structure
+//!
+//! The benchmarks are split into two separate modules for clear comparison:
+//!
+//! - **[`neo4j`]** - All Neo4j benchmarks using Cypher queries via HTTP API
+//! - **[`doublets`]** - All Doublets benchmarks using direct memory access
+//!
 //! ## Benchmarked Operations
 //!
 //! | Benchmark       | Operation                                      | What it measures                    |
@@ -18,8 +25,6 @@
 //!
 //! ## Storage Backends Tested
 //!
-//! Each benchmark runs against these 6 storage implementations:
-//!
 //! ### Doublets (4 variants)
 //! - `Doublets_United_Volatile` - In-memory unit storage
 //! - `Doublets_United_NonVolatile` - File-mapped unit storage
@@ -30,12 +35,25 @@
 //! - `Neo4j_NonTransaction` - Direct HTTP API calls
 //! - `Neo4j_Transaction` - Transaction wrapper (same underlying implementation)
 
-mod create;
-mod delete;
-mod each;
-mod update;
+pub mod doublets;
+pub mod neo4j;
 
-pub use create::create_links;
-pub use delete::delete_links;
-pub use each::*;
-pub use update::update_links;
+// Re-export all Neo4j benchmarks with neo4j_ prefix
+pub use neo4j::create_links as neo4j_create_links;
+pub use neo4j::delete_links as neo4j_delete_links;
+pub use neo4j::each_all as neo4j_each_all;
+pub use neo4j::each_concrete as neo4j_each_concrete;
+pub use neo4j::each_identity as neo4j_each_identity;
+pub use neo4j::each_incoming as neo4j_each_incoming;
+pub use neo4j::each_outgoing as neo4j_each_outgoing;
+pub use neo4j::update_links as neo4j_update_links;
+
+// Re-export all Doublets benchmarks with doublets_ prefix
+pub use self::doublets::create_links as doublets_create_links;
+pub use self::doublets::delete_links as doublets_delete_links;
+pub use self::doublets::each_all as doublets_each_all;
+pub use self::doublets::each_concrete as doublets_each_concrete;
+pub use self::doublets::each_identity as doublets_each_identity;
+pub use self::doublets::each_incoming as doublets_each_incoming;
+pub use self::doublets::each_outgoing as doublets_each_outgoing;
+pub use self::doublets::update_links as doublets_update_links;
